+++ misc/allow-block-list/src/lib.rs
@@ -231,7 +231,7 @@ where
         Ok(dummy::ConnectionHandler)
     }
 
-    fn on_swarm_event(&mut self, event: FromSwarm<Self::ConnectionHandler>) {
+    fn on_swarm_event(&mut self, event: FromSwarm) {
         match event {
             FromSwarm::ConnectionClosed(_) => {}
             FromSwarm::ConnectionEstablished(_) => {}