+++ protocols/mdns/src/behaviour.rs
@@ -224,7 +224,7 @@ where
         void::unreachable(ev)
     }
 
-    fn on_swarm_event(&mut self, event: FromSwarm<Self::ConnectionHandler>) {
+    fn on_swarm_event(&mut self, event: FromSwarm) {
         self.listen_addresses.on_swarm_event(&event);
 
         match event {