+++ protocols/perf/src/server.rs
@@ -118,7 +118,7 @@ impl NetworkBehaviour for Behaviour {
         )
     }
 
-    fn on_swarm_event(&mut self, event: FromSwarm<Self::ConnectionHandler>) {
+    fn on_swarm_event(&mut self, event: FromSwarm) {
         self.request_response.on_swarm_event(event);
     }